Min dator it is

+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/PlayerObject.h
@@ -13,11 +13,11 @@ public:
 	PlayerObject(Truck* truck, InputManager* input, sf::Sprite* sprite = nullptr, sf::Sprite* update = nullptr);
 	bool Update(float deltatime);
 	bool GetType();
-	int GetDamage();
+	float GetDamage();
 	sf::Vector2f GetDirection();
 private:
 	InputManager* m_input;
-	int m_damage;
+	float m_damage;
 	float m_cooldown;
 	sf::Vector2f m_direction;
 	sf::Sprite* m_update;
